placement of the (join) features in the right file (joinRequestImpl)+ adjust to work
