+++ Tournoi/src/main/java/be/TFTIC/Tournoi/bll/services/impl/ClanServiceImpl.java
@@ -73,47 +73,6 @@ public class ClanServiceImpl implements ClanService{
     }
 
     @Override
-    public JoinClanDTO joinClan(Long clanId, User user) {
-        Clan clan = getById(clanId);
-        Long userId= user.getId();
-
-        boolean userAlreadyMemberOfClan= clan.getMembers().stream()
-                .anyMatch(member-> member.getId().equals(userId));
-
-        if (userAlreadyMemberOfClan) {
-            return JoinClanDTO.fromEntity(clan, "You are already a member of this clan !");
-        }
-
-        Optional<JoinRequest> existingRequest = joinRequestRepository.findByUserAndClanAndStatus(user, clan, RequestStatus.PENDING);
-        //TODO passer par une méthode implémenter dans un service plutôt que le repo
-
-        if (existingRequest.isPresent()) {
-            return JoinClanDTO.fromEntity(clan, "You already have a pending join request for this clan.");
-        }
-
-        if(!clan.getIsPrivate()&&user.getRanking()>= clan.getMinimumTrophies()){
-            clan.getMembers().add(user);
-            clan.getRoles().put(user.getId(), ClanRole.MEMBER);
-            clanRepository.save(clan);
-            return JoinClanDTO.fromEntity(clan, "Welcome in " + clan.getName() + " !");
-
-        }
-
-        else if (clan.getIsPrivate()){
-            JoinRequest joinRequest= new JoinRequest();
-            joinRequest.setClan(clan);
-            joinRequest.setUser(user);
-            joinRequest.setStatus(RequestStatus.PENDING);
-            joinRequestRepository.save(joinRequest);
-            return JoinClanDTO.fromEntity(clan, "Clan "+ clan.getName() +" is private, your request is pending ");
-        }
-
-        else{
-            return JoinClanDTO.fromEntity(clan, "Cannot join this clan. Rank requirement not met ");
-        }
-    }
-
-    @Override
     public ClanDTO updateClan(User user,Long clanId, ClanFormEdit clanFormEdit) {
         Clan clan= getById(clanId);
 
@@ -201,32 +160,4 @@ public class ClanServiceImpl implements ClanService{
         return new MessageDTO("Role succesfully changed");
     }
 
-
-    @Override
-    public MessageDTO handleJoinRequest(Long clanId, Long userIdToChange, User currentUser, boolean accept) {
-        Clan clan = getById(clanId);
-
-        ClanRole currentUserRole = clan.getRoles().get(currentUser.getId());
-        if (currentUserRole == null ||
-                !(currentUserRole == ClanRole.PRESIDENT || currentUserRole == ClanRole.VICE_PRESIDENT)) {
-            return new MessageDTO("Only Presidents or Vice Presidents can handle join requests.");
-        }
-        User userToChange= getUserById(userIdToChange);
-
-        JoinRequest joinRequest= joinRequestRepository.findByUserAndClanAndStatus(userToChange,clan,RequestStatus.PENDING)
-                .orElseThrow(()-> new RuntimeException("Join request not found "));
-
-        if (accept) {
-            clan.getRoles().put(userToChange.getId(), ClanRole.MEMBER);
-            userRepository.save(userToChange);
-            joinRequest.setStatus(RequestStatus.APPROVED);
-            clanRepository.save(clan);
-            return new MessageDTO("Join request accepted, User has been added to the clan ");
-
-        } else {
-            joinRequest.setStatus(RequestStatus.REJECTED);
-            return new MessageDTO("Join request has been rejected.");
-        }
-
-    }
-}
+   }
